renamed jwt-core -> token-core, use central rule registry
