+++ extensions/common/iam/identity-trust/identity-trust-core/build.gradle.kts
@@ -8,7 +8,7 @@ dependencies {
     implementation(project(":spi:common:http-spi"))
     implementation(project(":spi:common:json-ld-spi"))
     implementation(project(":core:common:util"))
-    implementation(project(":core:common:jwt-core"))
+    implementation(project(":core:common:token-core"))
     implementation(project(":extensions:common:crypto:jws2020"))
     implementation(project(":extensions:common:crypto:jwt-verifiable-credentials"))
     implementation(project(":extensions:common:crypto:ldp-verifiable-credentials"))
