+++ extensions/common/iam/identity-trust/identity-trust-sts/identity-trust-sts-core/build.gradle.kts
@@ -8,7 +8,7 @@ dependencies {
     api(project(":spi:common:identity-trust-spi"))
     api(project(":spi:common:identity-trust-sts-spi"))
     implementation(project(":extensions:common:iam:identity-trust:identity-trust-sts:identity-trust-sts-embedded"))
-    implementation(project(":core:common:jwt-core"))
+    implementation(project(":core:common:token-core"))
 
     testImplementation(testFixtures(project(":spi:common:identity-trust-sts-spi")))
     testImplementation(project(":core:common:junit"))
