+++ extensions/common/iam/identity-trust/identity-trust-sts/identity-trust-sts-embedded/build.gradle.kts
@@ -10,7 +10,7 @@ dependencies {
     implementation(project(":core:common:util"))
     testImplementation(testFixtures(project(":spi:common:identity-trust-spi")))
     testImplementation(project(":core:common:junit"))
-    testImplementation(project(":core:common:jwt-core"))
+    testImplementation(project(":core:common:token-core"))
     testImplementation(libs.nimbus.jwt)
 }
