+++ extensions/common/iam/oauth2/oauth2-core/src/main/java/org/eclipse/edc/iam/oauth2/jwt/Oauth2JwtDecoratorRegistryRegistryImpl.java
@@ -15,7 +15,7 @@
 package org.eclipse.edc.iam.oauth2.jwt;
 
 import org.eclipse.edc.iam.oauth2.spi.Oauth2JwtDecoratorRegistry;
-import org.eclipse.edc.jwt.JwtDecoratorRegistryImpl;
+import org.eclipse.edc.token.JwtDecoratorRegistryImpl;
 
 /**
  * Registry for Oauth2 JWT decorators.
