+++ extensions/control-plane/transfer/transfer-data-plane/build.gradle.kts
@@ -29,7 +29,7 @@ dependencies {
     api(project(":spi:data-plane-selector:data-plane-selector-spi"))
 
     implementation(project(":extensions:common:api:control-api-configuration"))
-    implementation(project(":core:common:jwt-core"))
+    implementation(project(":core:common:token-core"))
 
     api(libs.jakarta.rsApi)
     api(libs.nimbus.jwt)
