+++ settings.gradle.kts
@@ -45,7 +45,7 @@ include(":core:common:boot")
 include(":core:common:connector-core")
 include(":core:common:jersey-providers")
 include(":core:common:junit")
-include(":core:common:jwt-core")
+include(":core:common:token-core")
 include(":core:common:policy-engine")
 include(":core:common:policy-evaluator")
 include(":core:common:state-machine")
