+++ spi/common/jwt-spi/src/main/java/org/eclipse/edc/jwt/spi/TokenValidationService.java
@@ -20,6 +20,9 @@ import org.eclipse.edc.spi.iam.TokenRepresentation;
 import org.eclipse.edc.spi.result.Result;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Arrays;
+import java.util.List;
+
 /**
  * Interface for validating token.
  */
@@ -33,7 +36,19 @@ public interface TokenValidationService {
      * @param rules               token validation rules that apply to the token
      * @return Result of the validation.
      */
-    Result<ClaimToken> validate(TokenRepresentation tokenRepresentation, PublicKeyResolver publicKeyResolver, TokenValidationRule... rules);
+    default Result<ClaimToken> validate(TokenRepresentation tokenRepresentation, PublicKeyResolver publicKeyResolver, TokenValidationRule... rules) {
+        return validate(tokenRepresentation, publicKeyResolver, Arrays.asList(rules));
+    }
+
+    /**
+     * Validates the token and offers possibility for additional information for validations.
+     *
+     * @param tokenRepresentation A token representation including the token to verify.
+     * @param publicKeyResolver   A {@link PublicKeyResolver} to obtain the public key with which to verify the token
+     * @param rules               token validation rules that apply to the token. Assume to be unmodifiable.
+     * @return Result of the validation.
+     */
+    Result<ClaimToken> validate(TokenRepresentation tokenRepresentation, PublicKeyResolver publicKeyResolver, List<TokenValidationRule> rules);
 
     /**
      * Validates the token.
