+++ system-tests/e2e-transfer-test/control-plane/build.gradle.kts
@@ -17,7 +17,7 @@ plugins {
 }
 
 dependencies {
-    implementation(project(":core:common:jwt-core"))
+    implementation(project(":core:common:token-core"))
     implementation(project(":core:control-plane:control-plane-core"))
     implementation(project(":data-protocols:dsp"))
     implementation(project(":extensions:common:vault:vault-filesystem"))
