+++ system-tests/sts-api/sts-api-test-runtime/build.gradle.kts
@@ -17,7 +17,7 @@ plugins {
 }
 
 dependencies {
-    implementation(project(":core:common:jwt-core"))
+    implementation(project(":core:common:token-core"))
     implementation(project(":extensions:common:http"))
     implementation(project(":extensions:common:iam:identity-trust:identity-trust-sts:identity-trust-sts-core"))
     implementation(project(":extensions:common:iam:identity-trust:identity-trust-sts:identity-trust-sts-api"))
